finished huge refactoring, implemented loaders
YAY. Scrooge and Goku are using the new character format. Megaman has
yet to be "fixed" in this way. Fixed the scale problem that made sprites
blurry when not being rendered at a 1.0 scale. Implemented rooms! Only
TomahawkMan's stage complies with the new stage format. Room transitions
are still to be implemented.